Changed Wording
I changed a bit of the wording to make the document easier to read. I think we should also elaborate on the answer for, "How do I know if I am sending too much?". The answer explains that the buffer for storing pending messages will grow, but doesn't explain if/why that is a problem.