Update 05-media-communication.md
Some sentence edits and added missing punctuation.

* wifi -> Wi-Fi
* twcc -> TWCC

I love the [history of how 1500 became the default size](https://blog.benjojo.co.uk/post/why-is-ethernet-mtu-1500), but that's way to in-depth to mention in this book.

I'm assuming "Media 101" is a WIP section, same with "Extension", "Mapping Payload Types to Codecs" and "Generic RTP Feedback"?