Heading Updates
The "Who is this book for." heading reads like a question without a question mark. I'd recommend changing it to either:
* Who is this book for?
or 
* Who this book is for: (what I've proposed here)

Additionally, the overall capitalization format for the headings doesn't feel consistent. I've proposed [sentence case](https://apastyle.apa.org/style-grammar-guidelines/capitalization/sentence-case) - where the first word is capitalized and all following words are lowercase.

If you want to adopt sentence case throughout the book, I'm happy to go through every page and make updates where necessary. Let me know!

I'd also recommend updating the title in the sidebar (WebRTC for the Curious) to match the overall title (WebRTC For The Curious) with a capital F in 'For' and a capital T in 'The'. Just to keep things consistent. I couldn't find where to propose this change though.